Fourier that should be working but ain't

+++ UnderwaterGLDemo/assets/shaders/surface.vert
@@ -1,6 +1,6 @@
 #version 430 core
 layout (location = 0) in vec3 position;
-layout (location = 1) in ivec2 texCoord;
+layout (location = 1) in vec2 texCoord;
 
 //uniform mat4 M;
 uniform mat4 V, invV;
